Add support for logging all headers received or sent
The logging directives currently allow logging of individual headers.
Sometimes, though, it can be useful to dump the full set ---
particularly when debugging. This change supports converting the
(almost) complete set of headers into a JSON-style list, and writing
this to the log. The escaping is likely to only work when used with
JSON-style logs.

Note that, due to h2o's internal handling of some headers, the output is
not a complete list. As a specific example, the Host header is not added
to h2o's internal headers list.