Allow logging of host (node) name
Adds a logging directive to log the hostname. As this is also referred
to as the nodename, and as both %h and %H are already taken, the code
selected is %N.

The hostname is read once, with the result cached for subsequent calls.

As the hostname uses only DNS characters, it does not require escaping.