Allow generation of request IDs
With this change, the headers module can generate a 128-bit random
request ID and add this to headers. The ID will be base32-encoded, which
offers a good balance of length and avoiding confusable characters. To
simplify reading of the IDs, it is split into five (almost) equal sized
blocks.

To enable ID generation, the header.setrequestid or
header.setrequestidifempty configuration keys can be used. Both take the
name of the header to use (e.g. 'X-Request-ID'), and differ in behaviour
when a matching header is already in the request. For front-end servers,
it is strongly recommended that the 'set' for is used, as this will
overwrite any ID header which ahs been passed in from outside. If there
are internal layers, though, the 'setifempty' variant may be more
appropriate as this will allow correlating logs from different systems.

When applied to a proxy (i.e. proxy.header.setrequestid), the header
will only be sent internally. When applied to the top-level headers
structure, the header will only be returned to the client. When both are
used together, the header will be both be sent to the proxy and returned
to the client, using the same value in each direction. The header's name
can be set differently (e.g. X-Trace-ID to the client, and X-Request-ID
to the proxy).