Avoid buffer overflow in zstd suport and change extension
When checking for precompressed files, the current code allocates a
buffer to take the path to a precompressed version of the main file. It
adds the length of '.gz' to the length of the path to do this.
Unfortunately, although this worked for gzip and brotli, zstd uses a
longer file extension. When support for zstd compressed files was added,
this was not updated to account for the longer extension. This creates a
buffer overflow (although due to heap padding this is almost certainly
safe -- but will not always be).

In addition, the official extension for zstd compressed files is 'zst',
rather than 'zstd', so update to allow this.